+++ src/KernelInterpolation.jl
@@ -14,6 +14,7 @@ module KernelInterpolation
 using DiffEqCallbacks: PeriodicCallback, PeriodicCallbackAffect
 using ForwardDiff: ForwardDiff
 using LinearAlgebra: Diagonal, Symmetric, I, norm, tr, dot, diagind, factorize
+using NearestNeighbors: KDTree, knn, inrange
 using Printf: @sprintf
 using Random: Random
 using ReadVTK: VTKFile, get_points, get_point_data, get_data